pyproject.toml
[tool.poetry]
name = "redis-benchmarks-specification"
version = "0.1.0"
description = ""
authors = ["filipecosta90 <filipecosta.90@gmail.com>"]
include = ["setups","test-suites"]

[tool.poetry.dependencies]
python = "^3.6"
Flask = "^2.0.1"
Werkzeug = "^2.0.1"
flask-restx = "^0.5.0"
redis = "^3.5.3"
marshmallow = "^3.12.2"
argparse = "^1.4.0"
Flask-HTTPAuth = "^4.4.0"
PyYAML = "^5.4.1"
docker = "^5.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
redis-benchmarks-spec-api = "redis_benchmarks_specification.api.api:main"
redis-benchmarks-spec-builder = "redis_benchmarks_specification.builder.builder:main"